Drop instrumentation-api-caching module and move weak cache implementation to instrumentation-api
