+++ conventions/src/main/kotlin/otel.javaagent-instrumentation.gradle.kts
@@ -12,5 +12,5 @@ base.archivesName.set(projectDir.parentFile.name)
 dependencies {
   // this only exists to make Intellij happy since it doesn't (currently at least) understand our
   // inclusion of this artifact inside of :instrumentation-api
-  compileOnly(project(":instrumentation-api-caching"))
+//  compileOnly(project(":instrumentation-api-caching"))
 }
