+++ conventions/src/main/kotlin/otel.library-instrumentation.gradle.kts
@@ -13,5 +13,5 @@ base.archivesName.set(projectDir.parentFile.name)
 dependencies {
   // this only exists to make Intellij happy since it doesn't (currently at least) understand our
   // inclusion of this artifact inside of :instrumentation-api
-  compileOnly(project(":instrumentation-api-caching"))
+//  compileOnly(project(":instrumentation-api-caching"))
 }
