+++ instrumentation/spring/spring-boot-autoconfigure/build.gradle.kts
@@ -55,7 +55,7 @@ dependencies {
 
   // this only exists to make Intellij happy since it doesn't (currently at least) understand our
   // inclusion of this artifact inside of :instrumentation-api
-  compileOnly(project(":instrumentation-api-caching"))
+//  compileOnly(project(":instrumentation-api-caching"))
 }
 
 tasks.compileTestJava {
