+++ javaagent-bootstrap/src/main/java/io/opentelemetry/javaagent/bootstrap/HelperResources.java
@@ -5,10 +5,10 @@
 
 package io.opentelemetry.javaagent.bootstrap;
 
-import io.opentelemetry.instrumentation.api.caching.Cache;
 import java.net.URL;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
+import io.opentelemetry.instrumentation.api.cache.Cache;
 
 /**
  * A holder of resources needed by instrumentation. We store them in the bootstrap classloader so
@@ -18,7 +18,7 @@ import java.util.concurrent.ConcurrentHashMap;
 public final class HelperResources {
 
   private static final Cache<ClassLoader, Map<String, URL>> RESOURCES =
-      Cache.builder().setWeakKeys().build();
+      Cache.builder().build();
 
   /** Registers the {@code payload} to be available to instrumentation at {@code path}. */
   public static void register(ClassLoader classLoader, String path, URL url) {
