+++ javaagent-extension-api/build.gradle.kts
@@ -19,5 +19,5 @@ dependencies {
 
   // this only exists to make Intellij happy since it doesn't (currently at least) understand our
   // inclusion of this artifact inside of :instrumentation-api
-  compileOnly(project(":instrumentation-api-caching"))
+//  compileOnly(project(":instrumentation-api-caching"))
 }
