+++ muzzle/build.gradle.kts
@@ -23,7 +23,7 @@ dependencies {
 
   // this only exists to make Intellij happy since it doesn't (currently at least) understand our
   // inclusion of this artifact inside of :instrumentation-api
-  compileOnly(project(":instrumentation-api-caching"))
+//  compileOnly(project(":instrumentation-api-caching"))
 
   testImplementation(project(":testing-common"))
   testImplementation("com.google.guava:guava")
