+++ muzzle/src/main/java/io/opentelemetry/javaagent/tooling/HelperInjector.java
@@ -5,7 +5,7 @@
 
 package io.opentelemetry.javaagent.tooling;
 
-import io.opentelemetry.instrumentation.api.caching.Cache;
+import io.opentelemetry.instrumentation.api.cache.Cache;
 import io.opentelemetry.javaagent.bootstrap.HelperResources;
 import io.opentelemetry.javaagent.tooling.muzzle.HelperResource;
 import java.io.File;
@@ -59,7 +59,7 @@ public class HelperInjector implements Transformer {
       };
 
   private static final Cache<Class<?>, Boolean> injectedClasses =
-      Cache.builder().setWeakKeys().build();
+      Cache.builder().build();
 
   private final String requestingName;
 
@@ -70,9 +70,9 @@ public class HelperInjector implements Transformer {
   private final Map<String, byte[]> dynamicTypeMap = new LinkedHashMap<>();
 
   private final Cache<ClassLoader, Boolean> injectedClassLoaders =
-      Cache.builder().setWeakKeys().build();
+      Cache.builder().build();
   private final Cache<ClassLoader, Boolean> resourcesInjectedClassLoaders =
-      Cache.builder().setWeakKeys().build();
+      Cache.builder().build();
 
   private final List<WeakReference<Object>> helperModules = new CopyOnWriteArrayList<>();
